pyproject.toml
[project]
name = "metadata_migrate_sync"
version = "0.0.1"
description = "metadata migrate and sync"
readme = "README.md"
authors = []
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.10.6",
    "typing_extensions>=4.12.2",
    "globus-sdk>=3.51.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.39",
    "pytest-datadir>=1.6.1",
]


[tool.uv]
dev-dependencies = [
    "pytest>=7.3.1",
    "pytest-datadir",
    "mypy>=1.15.0",
    "pytest-snapshot>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/metadata_migrate_sync"]


[tool.pytest.ini_options]
pythonpath = "."
addopts = [
    "--import-mode=importlib",
]

